rearrange treescript, add mypy
