+++ tox.ini
@@ -4,6 +4,8 @@ envlist =
     client-py37
     iscript-py36
     iscript-py37
+    treescript-py36
+    treescript-py37
 
 skipsdist = true
 
@@ -25,4 +27,12 @@ commands = tox -e py36
 
 [testenv:iscript-py37]
 changedir = {toxinidir}/iscript
+commands = tox -e py37,mypy
+
+[testenv:treescript-py36]
+changedir = {toxinidir}/treescript
+commands = tox -e py36
+
+[testenv:treescript-py37]
+changedir = {toxinidir}/treescript
 commands = tox -e py37,mypy
