+++ treescript/.coveragerc
@@ -1,3 +1,6 @@
 [run]
 branch = True
-omit = treescript/test/*, treescript/py2/robustcheckout.py
+omit =
+    tests/*
+    .tox/*
+    treescript/py2/robustcheckout.py
