+++ treescript/mypy.ini
@@ -1,0 +1,3 @@
+[mypy]
+python_version = 3.7
+ignore_missing_imports = True
