+++ treescript/setup.cfg
@@ -1,0 +1,33 @@
+[bdist_wheel]
+universal = 1
+
+[flake8]
+max-line-length = 140
+exclude = tests/*
+max-complexity=10
+
+[tool:pytest]
+norecursedirs =
+    .git
+    .tox
+    .env
+    dist
+    build
+python_files =
+    test_*.py
+    *_test.py
+    tests.py
+addopts =
+    -rxEfsw
+    --strict
+    --doctest-modules
+    --doctest-glob=\*.rst
+    --tb=short
+
+[options]
+package_dir=
+    =src
+packages=find:
+
+[options.packages.find]
+where=src
