+++ treescript/tests/test_versionmanip.py
@@ -4,12 +4,15 @@ import pytest
 from treescript.utils import DONTBUILD_MSG
 from scriptworker.context import Context
 from treescript.exceptions import TaskVerificationError
-from treescript.test import tmpdir, is_slice_in_list
 from treescript.script import get_default_config
 import treescript.versionmanip as vmanip
 
 
-assert tmpdir  # flake8
+def is_slice_in_list(s, l):
+    # Credit to https://stackoverflow.com/a/20789412/#answer-20789669
+    # With edits by Callek to be py3 and pep8 compat
+    len_s = len(s)  # so we don't recompute length of s on every iteration
+    return any(s == l[i:len_s + i] for i in range(len(l) - len_s + 1))
 
 
 @pytest.yield_fixture(scope='function')
