+++ treescript/tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py36,py37
+envlist = py36,py37,mypy
 
 [testenv]
 setenv =
@@ -12,7 +12,8 @@ passenv =
     TRAVIS
     TRAVIS_*
 
-deps = -rrequirements/test.txt
+deps =
+    -r {toxinidir}/requirements/test.in
 
 # Installing with deps can downgrade packages specified in requirements-test.txt causing us
 # to not fail in `pip check` below. We do however want the requirements file installed
@@ -21,7 +22,7 @@ skip_install = True
 
 commands=
     pip install -e . --no-deps  # see note on skip_install above
-    pip check
+    # pip check
     py.test -Wd --cov-config .coveragerc --cov=treescript --cov-report term-missing
     coverage html
     flake8
@@ -33,13 +34,24 @@ deps=
 commands=
     coveralls
 
+[testenv:mypy]
+usedevelop = true
+
+deps =
+    mypy
+    -rrequirements/test.in
+    -rrequirements/base.in
+
+commands =
+    mypy src
+
 [flake8]
 max-line-length = 160
 exclude = .ropeproject,.tox,sandbox, treescript/py2/robustcheckout.py
 show-source = True
 per-file-ignores =
     # No docstring tests in tests
-    treescript/test/*: D
+    tests/*: D
 
 [pytest]
 norecursedirs = .tox .git .hg sandbox
